[fix] test 코드 분리
