+++ src/utils/validate.js
@@ -7,7 +7,7 @@ import {
 } from "../data/constant.js";
 
 export const validateInputPriceType = (price) => {
-  if (!price || isNaN(price) || price.includes(" ")) {
+  if (!price || isNaN(price) || String(price).includes(" ")) {
     throw new Error("금액은 숫자만 입력 가능하며 빈 값은 허용되지 않습니다.");
   }
 };
